Fix #4: ListView broken inside of Hub.Section
The ListView wasn't rendering properly when used inside of a
Hub.Section or Pivot.Item. The problem was the ListView was
being instantiated in a tick of the event loop that was earlier
than the one that placed it into the DOM. This was due to the
Hub.Section/Pivot.Item mounting a React component inside of it
before it was in the DOM.

The fix is for the Hub.Section/Pivot.Item to mount React
components inside of it only after it is in the DOM. This is done
by taking advantage of the HubSection/PivotItem's
isDeclarativeControlContainer hook.